style: tree view refactor
Cleanup and prep for future changes
